requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
cryptography==41.0.7
python-multipart==0.0.6
email-validator==2.1.0